updated vhosts

+++ pyserv.py
@@ -776,13 +776,20 @@ class Services:
 					self.msg(source, "Syntax: AUTH <username> <password>")
 			elif text.lower().split()[0] == "vhost" and self.auth(source) != 0:
 				if len(text.split()) == 2:
-					self.query("delete from vhosts where user = '%s'" % self.auth(source))
-					self.query("insert into vhosts values ('%s','%s','0')" % (self.auth(source), text.split()[1]))
-					self.msg(source, "Your new vhost\2 %s\2 has been requested" % text.split()[1])
-					for data in self.query("select host from online where uid = '%s'" % source):
-						self.send(":%s CHGHOST %s %s" % (self.bot, source, data[0]))
-					for data in self.query("select uid from opers"):
-						self.omsg(data[0], "vHost request received from\2 %s\2" % self.auth(source))
+					if len(arg[1]) < 5:
+						self.msg(source, "Your vhost is too short.")
+					elif arg[1].find(".") == -1:
+						self.msg(source, "Invalid vhost. Where's the dot?")
+					elif arg[1][-2] == "." or arg[1][-1] == ".":
+						self.msg(source, "Domain ending is too short.")
+					else:
+						self.query("delete from vhosts where user = '%s'" % self.auth(source))
+						self.query("insert into vhosts values ('%s','%s','0')" % (self.auth(source), text.split()[1]))
+						self.msg(source, "Your new vhost\2 %s\2 has been requested" % text.split()[1])
+						for data in self.query("select host from online where uid = '%s'" % source):
+							self.send(":%s CHGHOST %s %s" % (self.bot, source, data[0]))
+						for data in self.query("select uid from opers"):
+							self.omsg(data[0], "vHost request received from\2 %s\2" % self.auth(source))
 				elif len(arg) == 1:
 					self.query("delete from vhosts where user = '%s'" % self.auth(source))
 					self.msg(source, "Done.")
